14/12/2022 -> ReciclerView agafa dades de Firebase ok + repetible

+++ app/src/main/java/com/example/zenhabit/Fragments/TasksFragment.kt
@@ -12,12 +12,13 @@ import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.fragment.findNavController
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import com.example.zenhabit.R
-import com.example.zenhabit.adapter.AdapterTasques
+import com.example.zenhabit.adapter.AdapterObjectius
 import com.example.zenhabit.databinding.FragmentTasksBinding
 import com.example.zenhabit.models.Habit
+import com.example.zenhabit.models.Objectius
 import com.example.zenhabit.models.Repte
-import com.example.zenhabit.models.Tasca
 import com.facebook.shimmer.ShimmerFrameLayout
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.firestore.FirebaseFirestore
@@ -29,8 +30,10 @@ class TasksFragment : Fragment() {
     private lateinit var _binding: FragmentTasksBinding
     private val binding get() = _binding
 
-    private lateinit var data: MutableList<Tasca>
-    private lateinit var mAdapter: AdapterTasques
+    val db = FirebaseFirestore.getInstance()
+    private lateinit var data: MutableList<Objectius>
+    private lateinit var mRecyclerView: RecyclerView
+    private lateinit var mAdapter: AdapterObjectius
     private lateinit var shimmerFrameLayout: ShimmerFrameLayout
     private lateinit var shimmerFrameLayoutObjDiari: ShimmerFrameLayout
 
@@ -87,36 +90,53 @@ class TasksFragment : Fragment() {
 
 
 
-
-//      RecyclerView shimmer
-        val mRecyclerView = binding.rvTasques
+//----------------NEW RECYCLERVIEW-----------------
+//cargar shimmer
+        mRecyclerView = binding.rvTasques
         val mLayoutManager = LinearLayoutManager(this.getActivity())
-
         shimmerFrameLayout = binding.shimmer
         shimmerFrameLayout.startShimmer()
-
-//        cargar
-        Handler(Looper.getMainLooper()).postDelayed({
-            binding.rvTasques.visibility = View.VISIBLE
-            shimmerFrameLayout.stopShimmer()
-            shimmerFrameLayout.visibility = View.INVISIBLE
-        }, 1000)
-
-
-//        cargar recyclerview
-        mRecyclerView.setLayoutManager(mLayoutManager);
-        data = dataInicialize() as MutableList<Tasca>
-        mAdapter = AdapterTasques(
-            data,
-            { index -> deleteItem(index) },
-            { nom, hora -> sendItem(nom, hora) });
-        mRecyclerView.setAdapter(mAdapter)
-
-
+//cargar recyclerview
+        mRecyclerView.layoutManager = mLayoutManager
+        loadData()
 
         return view
     }
 
+    private fun loadData() {
+
+        var ret: ArrayList<Objectius> = ArrayList()
+
+        val docref = db.collection("Usuaris").document(Firebase.auth.currentUser!!.uid)
+        docref.get().addOnSuccessListener { document ->
+            if (document != null) {
+
+                ret = Objectius.dataFirebaseToObjectius(document)
+
+//shimmer desaparece
+                binding.rvTasques.visibility = View.VISIBLE
+                shimmerFrameLayout.stopShimmer()
+                shimmerFrameLayout.visibility = View.INVISIBLE
+
+
+                mAdapter = AdapterObjectius(
+                    ret,
+                    { index -> deleteItem(index) },
+                    { nom, hora -> sendItem(nom, hora) });
+
+                mRecyclerView.setAdapter(mAdapter)
+
+            } else {
+                //ERROR
+                Log.d("TAG", "DocumentSnapshot data: NO SE ENCONTRO EL DOCUMENTO")
+
+            }
+
+        }.addOnFailureListener { exception ->
+            Log.d("TAG", "ERROR AL OBTENER ${exception}")
+
+        }
+    }
     private fun sendItem(nom: String, hora: String) {
         val action =
             TasksFragmentDirections.actionTasksFragment2ToCreateEditTaskFragment(nom, hora)
@@ -135,29 +155,6 @@ class TasksFragment : Fragment() {
 
     }
 
-    private fun dataInicialize(): ArrayList<Tasca> {
 
-        val tasquesList: ArrayList<Tasca> = ArrayList()
-
-
-        tasquesList.add(Tasca("Llegir", "10:00", "Tasca"))
-        tasquesList.add(Tasca("Caminar", "12:00", "Habit"))
-        tasquesList.add(Tasca("Comprar", "22:00", "Tasca"))
-        tasquesList.add(Tasca("Llegir", "10:00", "Tasca"))
-        tasquesList.add(Tasca("Caminar", "12:00", "Habit"))
-        tasquesList.add(Tasca("Comprar", "22:00", "Tasca"))
-        tasquesList.add(Tasca("Llegir", "10:00", "Tasca"))
-        tasquesList.add(Tasca("Caminar", "12:00", "Habit"))
-        tasquesList.add(Tasca("Comprar", "22:00", "Tasca"))
-        tasquesList.add(Tasca("Llegir", "10:00", "Tasca"))
-        tasquesList.add(Tasca("Caminar", "12:00", "Habit"))
-        tasquesList.add(Tasca("Comprar", "22:00", "Tasca"))
-        tasquesList.add(Tasca("Llegir", "10:00", "Tasca"))
-        tasquesList.add(Tasca("Caminar", "12:00", "Habit"))
-        tasquesList.add(Tasca("Comprar", "22:00", "Tasca"))
-
-        return tasquesList
-
-    }
 
 }
